+++ comb_spec_searcher/strategies/rule.py
@@ -60,6 +60,7 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
         ] = (None)
         self.subsamplers: Optional[Tuple[Callable[..., CombinatorialObjectType], ...]]
         self._children = children
+        self._non_empty_children: Optional[Tuple[CombinatorialClassType, ...]] = None
 
     @property
     def strategy(
@@ -148,7 +149,11 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
         Return the tuple of non-empty combinatorial classes that are found
         by applying the decomposition function.
         """
-        return tuple(child for child in self.children if not child.is_empty())
+        if self._non_empty_children is None:
+            self._non_empty_children = tuple(
+                child for child in self.children if not child.is_empty()
+            )
+        return self._non_empty_children
 
     @abc.abstractmethod
     def count_objects_of_size(self, n: int, **parameters: int) -> int:
@@ -194,11 +199,17 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
         def brute_force_count(
             comb_class: CombinatorialClassType, n: int, **parameters
         ) -> int:
+            assert set(comb_class.extra_parameters) == set(
+                parameters
+            ), f"{comb_class.extra_parameters, set(parameters)}"
             return len(list(comb_class.objects_of_size(n, **parameters)))
 
         def brute_force_generation(
             comb_class: CombinatorialClassType, n: int, **parameters: int
         ) -> Iterator[CombinatorialObjectType]:
+            assert set(comb_class.extra_parameters) == set(
+                parameters
+            ), f"{comb_class.extra_parameters, set(parameters)}"
             yield from comb_class.objects_of_size(n, **parameters)
 
         actual_count = brute_force_count(self.comb_class, n, **parameters)
@@ -331,7 +342,7 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
         return self._constructor
 
     def is_equivalence(self):
-        return len(self.non_empty_children()) == 1 and self.constructor.is_equivalence()
+        return self.strategy.can_be_equivalent() and len(self.non_empty_children()) == 1
 
     def backward_map(
         self, objs: Tuple[Optional[CombinatorialObjectType], ...]
@@ -390,9 +401,12 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
             assert (
                 self.subrecs is not None
             ), "you must call the set_subrecs function first"
-            res = self.constructor.get_recurrence(self.subrecs, n, **parameters)
+            try:
+                res = self.constructor.get_recurrence(self.subrecs, n, **parameters)
+            except AssertionError:
+                raise ValueError(f"issue with rule:\n {self}")
             self.count_cache[key] = res
-        #     # THE FOLLOWING CODE SNIPPET IS FOR DEBUGGING PURPOSES
+            # THE FOLLOWING CODE SNIPPET IS FOR DEBUGGING PURPOSES
         #     if self.comb_class.extra_parameters:
         #         print(self)
         #         print("n =", n, parameters)
@@ -421,6 +435,8 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
         #         for string in fusion_attrs:
         #             if hasattr(self.constructor, string):
         #                 print(string + ":", getattr(self.constructor, string))
+        #         if isinstance(self.constructor, CartesianProduct):
+        #             print(list(self.constructor._valid_compositions(n, **parameters)))
         #     print("result:", res)
         # assert res == len(list(self.comb_class.objects_of_size(n, **parameters))), (
         #     "counting failed for the rule \n{}\nparameters: n = {}, {}\n"
@@ -557,8 +573,6 @@ class EquivalencePathRule(Rule[CombinatorialClassType, CombinatorialObjectType])
     @property
     def constructor(self) -> DisjointUnion:
         if self._constructor is None:
-            if not self.comb_class.extra_parameters:
-                return DisjointUnion(self.comb_class, self.children)
             extra_parameters: Dict[str, str] = {
                 k: k for k in self.comb_class.extra_parameters
             }
@@ -571,8 +585,13 @@ class EquivalencePathRule(Rule[CombinatorialClassType, CombinatorialObjectType])
                     for parent_var, child_var in extra_parameters.items()
                     if child_var in rules_parameters
                 }
+            fixed_values = {
+                k: 0
+                for k in self.children[0].extra_parameters
+                if k not in extra_parameters.values()
+            }
             self._constructor = DisjointUnion(
-                self.comb_class, self.children, (extra_parameters,)
+                self.comb_class, self.children, (extra_parameters,), (fixed_values,)
             )
         return self._constructor
 
@@ -674,8 +693,13 @@ class ReverseRule(Rule[CombinatorialClassType, CombinatorialObjectType]):
             flipped_extra_params = {
                 b: a for a, b in original_constructor.extra_parameters[0].items()
             }
+            fixed_values = {
+                k: 0
+                for k in self.children[0].extra_parameters
+                if k not in flipped_extra_params.values()
+            }
             self._constructor = DisjointUnion(
-                self.comb_class, self.children, (flipped_extra_params,)
+                self.comb_class, self.children, (flipped_extra_params,), (fixed_values,)
             )
         return self._constructor