+++ comb_spec_searcher/strategies/strategy.py
@@ -165,6 +165,13 @@ class AbstractStrategy(
         return self._workable
 
     @abc.abstractmethod
+    def can_be_equivalent(self) -> bool:
+        """
+        Return True if every Rule returned with one non-empty child is an
+        equivalence rule.
+        """
+
+    @abc.abstractmethod
     def decomposition_function(
         self, comb_class: CombinatorialClassType
     ) -> Optional[Tuple[CombinatorialClassType, ...]]:
@@ -362,6 +369,10 @@ class CartesianProductStrategy(
             workable=workable,
         )
 
+    @staticmethod
+    def can_be_equivalent() -> bool:
+        return True
+
     def constructor(
         self,
         comb_class: CombinatorialClassType,
@@ -372,7 +383,9 @@ class CartesianProductStrategy(
             if children is None:
                 raise StrategyDoesNotApply("Strategy does not apply")
         return CartesianProduct(
-            children, extra_parameters=self.extra_parameters(comb_class, children)
+            comb_class,
+            children,
+            extra_parameters=self.extra_parameters(comb_class, children),
         )
 
 
@@ -398,6 +411,10 @@ class DisjointUnionStrategy(Strategy[CombinatorialClassType, CombinatorialObject
             possibly_empty=possibly_empty,
             workable=workable,
         )
+
+    @staticmethod
+    def can_be_equivalent() -> bool:
+        return True
 
     def constructor(
         self,
@@ -505,6 +522,10 @@ class VerificationStrategy(
         if children is None:
             children = self.decomposition_function(comb_class)
         return VerificationRule(self, comb_class, children)
+
+    @staticmethod
+    def can_be_equivalent() -> bool:
+        return False
 
     def pack(self, comb_class: CombinatorialClassType) -> "StrategyPack":
         """